Issues fixed in tests
