+++ delta_backend/Makefile
@@ -8,4 +8,16 @@ package: build
 test:
 	python -m unittest
 
+check-conversion:
+	python tests/check_conversion.py
+
+coverage-run:
+	coverage run -m unittest discover -v
+
+coverage-report:
+	coverage report -m 
+
+coverage-html:
+	coverage html  
+
 .PHONY: build package
