+++ delta_backend/pyproject.toml
@@ -14,6 +14,7 @@ boto3 = "~1.26.90"
 mypy-boto3-dynamodb = "^1.26.164"
 moto = "~4.2.11"
 python-stdnum = "^1.20"
+coverage = "^7.8.0"
 
 [tool.poetry.group.dev.dependencies]
 coverage = "^7.8.0"
